requirements.txt
tensorflow
streamlit
gdown
numpy
Pillow

langchain
langchain-groq
python-dotenv
langchain-community
sentence-transformers
tf-keras
chromadb

streamlit
tensorflow
